Implement defaulting for MaxInputPercent

+++ BrickController2/BrickController2/BusinessLogic/PlayLogic.cs
@@ -235,7 +235,7 @@ namespace BrickController2.BusinessLogic
                 }
             }
 
-            axisValue = controllerAction.MaxInputPercent > 0 ? axisValue * 100F / controllerAction.MaxInputPercent : 0;
+            axisValue = controllerAction.MaxInputPercent > 0 ? axisValue * 100F / (float)controllerAction.MaxInputPercent : 0;
             
             if (controllerAction.AxisCharacteristic == ControllerAxisCharacteristic.Exponential)
             {
